refactor: extract code finisher from ruby-generator index
Move finish, initTargets, and finishTargets methods from index.js to
code-finisher.js. index.js is now 179 lines (was 844).

Part of #302
