+++ packages/scratch-gui/src/lib/ruby-generator/code-finisher.js
@@ -1,0 +1,157 @@
+// === Smalruby: This file is Smalruby-specific (code finishing/post-processing for RubyGenerator) ===
+
+/**
+ * Register code finishing methods on the RubyGenerator.
+ * Handles: finish, initTargets, finishTargets.
+ * @param {object} Generator - The RubyGenerator instance.
+ * @returns {object} The Generator instance.
+ */
+export default function (Generator) {
+    Generator.finish = function (code, options) {
+        const defs = [];
+        for (const name in this.definitions_) {
+            const def = this.definitions_[name];
+            if (this.isString(def)) {
+                if (name.match(/^require__/)) {
+                    this.requires_[name] = def;
+                } else if (name.match(/^prepare__/)) {
+                    this.prepares_[name] = def;
+                } else {
+                    defs.push(def);
+                }
+            }
+        }
+
+        const comments = Generator.getTargetCommentTexts();
+
+        // Detect @ruby:class comments
+        let classComment = null;
+        const otherComments = [];
+        for (const comment of comments) {
+            if (comment === '@ruby:class' || comment.startsWith('@ruby:class:')) {
+                classComment = comment;
+            } else {
+                otherComments.push(comment);
+            }
+        }
+
+        // Generate module...end blocks from collected module method codes
+        let moduleCode = '';
+        if (classComment) {
+            // Parse include= from class comment to determine module order
+            const includeModuleNames = [];
+            if (classComment.startsWith('@ruby:class:')) {
+                const attrPart = classComment.slice('@ruby:class:'.length);
+                const attrs = attrPart.split(',');
+                for (const attr of attrs) {
+                    const includeMatch = attr.match(/^include=(.+)$/);
+                    if (includeMatch) {
+                        includeModuleNames.push(includeMatch[1]);
+                    }
+                }
+            }
+
+            // Generate module blocks in include order
+            for (const moduleName of includeModuleNames) {
+                const methods = this._moduleMethodCodes[moduleName];
+                if (methods && methods.length > 0) {
+                    const methodsCode = methods.join('\n');
+                    moduleCode += `module ${moduleName}\n`;
+                    moduleCode += this.prefixLines(methodsCode, this.INDENT);
+                    moduleCode += `end\n\n`;
+                }
+            }
+        }
+
+        // For version 1 file output (withSpriteNew), use Sprite.new format
+        // even when @ruby:class comment is present.
+        // For version 2, @ruby:class takes priority over withSpriteNew.
+        // For version 2 stage targets without @ruby:class, auto-wrap with class Stage.
+        if (classComment && this.version !== '1') {
+            code = this._wrapWithClass(
+                code, classComment, options && options.withSpriteNew
+            );
+        } else if (this.version !== '1' && options && options.withSpriteNew) {
+            // Version 2: auto-wrap with class (both sprite and stage)
+            if (code.length > 0) {
+                code = this._wrapWithClass(code, '@ruby:class', true);
+            }
+        } else if (options && options.withSpriteNew) {
+            const spriteNewCode = this.spriteNew(this.currentTarget);
+            if (code.length > 0) {
+                code = this.prefixLines(code, this.INDENT);
+            }
+            code = `${spriteNewCode} do\n${code}end\n`;
+        }
+
+        // Add non-class target comments AFTER class wrapping so they appear
+        // before the class definition, not inside it.
+        if (otherComments.length > 0) {
+            const commentCodes = otherComments.map(comment => `${this.prefixLines(comment, '# ')}\n`);
+            code = `${commentCodes.join('\n')}\n${code}`;
+        }
+
+        if (defs.length === 0 && moduleCode.length === 0 && code.length === 0) {
+            return '';
+        }
+
+        let s = '';
+        if (defs.length > 0) {
+            s += `${defs.join('\n')}\n\n`;
+        }
+
+        return s + moduleCode + code;
+    };
+
+    Generator.initTargets = function (options) {
+        this.requires_ = {};
+        this.prepares_ = {};
+
+        if (options && Object.prototype.hasOwnProperty.call(options, 'requires')) {
+            options.requires.forEach(name => {
+                this.requires_[`require__${name}`] = `require "${name}"`;
+            });
+        }
+    };
+
+    Generator.finishTargets = function (code, _options) {
+        let s = '';
+        const requires = Object.keys(this.requires_).map(name => this.requires_[name]);
+        if (requires.length > 0) {
+            s += `${requires.join('\n')}\n\n`;
+        }
+
+        const prepares = Object.keys(this.prepares_).map(name => this.prepares_[name]);
+        if (prepares.length > 0) {
+            s += `${prepares.join('\n')}\n\n`;
+        }
+
+        // Deduplicate module definitions in multi-target output.
+        // Extract all module...end blocks, keep unique ones, place them before class definitions.
+        const moduleRegex = /^module (\w+)\n[\s\S]*?^end\n/gm;
+        const seenModules = new Set();
+        const uniqueModules = [];
+        let match;
+        while ((match = moduleRegex.exec(code)) !== null) {
+            const moduleName = match[1];
+            if (!seenModules.has(moduleName)) {
+                seenModules.add(moduleName);
+                uniqueModules.push(match[0]);
+            }
+        }
+
+        if (uniqueModules.length > 0) {
+            // Remove all module definitions from code
+            code = code.replace(moduleRegex, '');
+            // Clean up extra blank lines left by removal
+            code = code.replace(/\n{3,}/g, '\n\n').replace(/^\n+/, '');
+            // Prepend unique modules
+            const modulesCode = uniqueModules.join('\n');
+            code = `${modulesCode}\n${code}`;
+        }
+
+        return s + code;
+    };
+
+    return Generator;
+}
